+++ Assets/__Project/Scripts/Platform/Visual/PlatformVisual.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Combat.Battlefield;
 using UnityEngine;
 
 namespace Platform
@@ -7,8 +8,9 @@ namespace Platform
     {
         public Vector3 Position { get; set; }
         public Vector2 Size { get; set; }
+        public PlatformHexSurface Surface { get; set; }
         public List<Vector3> TopBoundary { get; set; } = new();
-        public GameObject GameObject { get; set; } 
+        public GameObject GameObject { get; set; }
         
         public void UpdateVisual()
         {